Rename web.model package
package ru.mystamps.web.model renamed to ru.mystamps.web.controller.dto
