+++ src/main/java/ru/mystamps/web/controller/SeriesController.java
@@ -59,13 +59,13 @@ import ru.mystamps.web.Url;
 import ru.mystamps.web.controller.converter.annotation.Category;
 import ru.mystamps.web.controller.converter.annotation.Country;
 import ru.mystamps.web.controller.converter.annotation.CurrentUser;
+import ru.mystamps.web.controller.dto.AddImageForm;
+import ru.mystamps.web.controller.dto.AddSeriesForm;
+import ru.mystamps.web.controller.dto.AddSeriesSalesForm;
 import ru.mystamps.web.dao.dto.EntityWithIdDto;
 import ru.mystamps.web.dao.dto.LinkEntityDto;
 import ru.mystamps.web.dao.dto.PurchaseAndSaleDto;
 import ru.mystamps.web.dao.dto.SeriesInfoDto;
-import ru.mystamps.web.model.AddImageForm;
-import ru.mystamps.web.model.AddSeriesForm;
-import ru.mystamps.web.model.AddSeriesSalesForm;
 import ru.mystamps.web.service.CategoryService;
 import ru.mystamps.web.service.CollectionService;
 import ru.mystamps.web.service.CountryService;
