+++ src/main/java/ru/mystamps/web/controller/dto/ActivateAccountForm.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package ru.mystamps.web.model;
+package ru.mystamps.web.controller.dto;
 
 import javax.validation.GroupSequence;
 import javax.validation.constraints.Pattern;
