+++ src/main/java/ru/mystamps/web/controller/dto/AddParticipantForm.java
@@ -15,34 +15,35 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package ru.mystamps.web.model;
+package ru.mystamps.web.controller.dto;
 
-import javax.validation.GroupSequence;
 import javax.validation.constraints.Size;
 
 import org.hibernate.validator.constraints.NotEmpty;
+import org.hibernate.validator.constraints.URL;
 
 import lombok.Getter;
 import lombok.Setter;
 
-import ru.mystamps.web.service.dto.RegisterAccountDto;
-import ru.mystamps.web.validation.jsr303.Email;
+import ru.mystamps.web.service.dto.AddParticipantDto;
 
-import static ru.mystamps.web.validation.ValidationRules.EMAIL_MAX_LENGTH;
+import static ru.mystamps.web.validation.ValidationRules.PARTICIPANT_NAME_MAX_LENGTH;
+import static ru.mystamps.web.validation.ValidationRules.PARTICIPANT_NAME_MIN_LENGTH;
+import static ru.mystamps.web.validation.ValidationRules.PARTICIPANT_URL_MAX_LENGTH;
 
 @Getter
 @Setter
-@GroupSequence({
-	RegisterAccountForm.class,
-	Group.Level1.class,
-	Group.Level2.class,
-	Group.Level3.class
-})
-public class RegisterAccountForm implements RegisterAccountDto {
+public class AddParticipantForm implements AddParticipantDto {
 	
-	@NotEmpty(groups = Group.Level1.class)
-	@Size(max = EMAIL_MAX_LENGTH, message = "{value.too-long}", groups = Group.Level2.class)
-	@Email(groups = Group.Level3.class)
-	private String email;
+	@NotEmpty
+	@Size.List({
+		@Size(min = PARTICIPANT_NAME_MIN_LENGTH, message = "{value.too-short}"),
+		@Size(max = PARTICIPANT_NAME_MAX_LENGTH, message = "{value.too-long}")
+	})
+	private String name;
+	
+	@URL
+	@Size(max = PARTICIPANT_URL_MAX_LENGTH, message = "{value.too-long}")
+	private String url;
 	
 }
