+++ src/test/groovy/ru/mystamps/web/service/SeriesSalesServiceImplTest.groovy
@@ -23,7 +23,7 @@ import spock.lang.Unroll
 import ru.mystamps.web.dao.SeriesSalesDao
 import ru.mystamps.web.dao.dto.AddSeriesSalesDbDto
 import ru.mystamps.web.dao.dto.Currency
-import ru.mystamps.web.model.AddSeriesSalesForm
+import ru.mystamps.web.controller.dto.AddSeriesSalesForm
 import ru.mystamps.web.tests.DateUtils
 
 @SuppressWarnings(['ClassJavadoc', 'MethodName', 'NoDef', 'NoTabCharacter', 'TrailingWhitespace'])
