+++ src/test/groovy/ru/mystamps/web/service/TransactionParticipantServiceImplTest.groovy
@@ -22,7 +22,7 @@ import spock.lang.Specification
 import ru.mystamps.web.dao.TransactionParticipantDao
 import ru.mystamps.web.dao.dto.AddParticipantDbDto
 import ru.mystamps.web.dao.dto.EntityWithIdDto
-import ru.mystamps.web.model.AddParticipantForm
+import ru.mystamps.web.controller.dto.AddParticipantForm
 
 @SuppressWarnings(['ClassJavadoc', 'MethodName', 'NoDef', 'NoTabCharacter', 'TrailingWhitespace'])
 class TransactionParticipantServiceImplTest extends Specification {
