+++ src/test/groovy/ru/mystamps/web/service/UserServiceImplTest.groovy
@@ -25,7 +25,7 @@ import ru.mystamps.web.dao.UserDao
 import ru.mystamps.web.dao.dto.AddUserDbDto
 import ru.mystamps.web.dao.dto.UserDetails
 import ru.mystamps.web.dao.dto.UsersActivationDto
-import ru.mystamps.web.model.ActivateAccountForm
+import ru.mystamps.web.controller.dto.ActivateAccountForm
 import ru.mystamps.web.tests.DateUtils
 
 @SuppressWarnings(['ClassJavadoc', 'MethodName', 'NoDef', 'NoTabCharacter', 'TrailingWhitespace'])
