+++ src/test/groovy/ru/mystamps/web/service/UsersActivationServiceImplTest.groovy
@@ -24,7 +24,7 @@ import ru.mystamps.web.dao.UsersActivationDao
 import ru.mystamps.web.dao.dto.AddUsersActivationDbDto
 import ru.mystamps.web.dao.dto.UsersActivationDto
 import ru.mystamps.web.dao.dto.UsersActivationFullDto
-import ru.mystamps.web.model.RegisterAccountForm
+import ru.mystamps.web.controller.dto.RegisterAccountForm
 import ru.mystamps.web.service.dto.SendUsersActivationDto
 import ru.mystamps.web.tests.DateUtils
 import ru.mystamps.web.validation.ValidationRules
